# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API",
        "description": "Documentation de l'API e-commerce.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Lister les commandes (user: les siennes, admin: toutes)",
                "description": "Display a listing of the resource.",
                "operationId": "8f3183bfa0ecb8a2c864cbf972b8061d",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Créer une commande",
                "description": "Store a newly created resource in storage.",
                "operationId": "b6490f9d9026dbe871b41207e118fffb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "shipping_address",
                                    "phone"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "shipping_address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créée"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Afficher une commande",
                "description": "Display the specified resource.",
                "operationId": "3a2d78e76e9fc3f06bb879fb7d48b40d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Mettre à jour le statut (admin)",
                "description": "Update the specified resource in storage.",
                "operationId": "bcc4793b40baabcf0d2bcf39e5705b9a",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "confirmed",
                                            "shipped",
                                            "delivered",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Annuler (user) ou supprimer (admin) la commande",
                "description": "Remove the specified resource from storage.",
                "operationId": "94988f9d23bb384ab6ee333a58ebe103",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "0": {
                            "sanctum": []
                        }
                    }
                ]
            }
        },
...
    },
...
}